pyproject.toml
[tool.poetry]
name = "SeaPlayer"
version = "0.10.0.dev27"
description = "SeaPlayer is a player that works in the terminal."
authors = ["Romanin <60302782+romanin-rf@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
#homepage = "https://romanin-rf.github.io/SeaPlayer"
repository = "https://github.com/romanin-rf/SeaPlayer"
keywords = [
    "textual", "tui",
    "seaplayer", "player",
    "audio", "wav", "mp3", "ogg", "flac"
]
classifiers = [
    "Environment :: Console",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS"
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
textual = ">=1.0.0"
rich = ">=13.3.3"
rich-click = "^1.8.5"
pillow = "^11.0.0"
numpy = "^2.2.1"
properties-py = "^1.2.1"
pyyaml = "^6.0.2"
platformdirs = "^4.3.6"
typing-extensions = "^4.12.2"
pydantic = "^2.10.4"
seaplayer-audio = {git = "https://github.com/romanin-rf/seaplayer-audio", rev = "dev"}

[tool.poetry.group.dev.dependencies]
pyinstaller-builder = {git = "https://git.smokeofanarchy.ru/Romanin/pyinstaller-builder.git"}
rich = "^13.9.4"
mkdocs = "^1.6.1"
mkdocs-exclude = "^1.0.2"
mkdocs-autorefs = "^1.2.0"
mkdocs-rss-plugin = "^1.17.1"
mkdocs-mermaid2-plugin = "^1.2.1"
mkdocstrings = {extras = ["crystal", "python"], version = "^0.27.0"}
mkdocstrings-python = ">=0.18"
mkdocs-material = "^9.5.49"
mkdocs-material-extensions = "^1.3.1"

[tool.poetry.scripts]
seaplayer = 'seaplayer.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
